refactor: streamline create & edit account pages (#849)
* 833 move components for nextjs routing

* add missing routes

* refactor: Extract ManageAccountForm.tsx

* refactor: create & edit account pages to use ManageAccountForm

* resolve conflicts

* refactor: split ManageAccountContainer into ManageAccountForm and page level code

* feat: different background images on edit account page than create account page

* feat: improve responsive background images on accounts pages

* feat: improve padding on edit account form

* fix: load birthdate correctly

* fix: bad sponsor redirect
